gpllib: bootloader id should be uint16_t
The id is of the form 0xABCD: use uint16_t instead of int.

+++ com32/gplinclude/disk/bootloaders.h
@@ -9,6 +9,9 @@
 
 #ifndef __BOOTLOADERS_H_
 #define __BOOTLOADERS_H_
-void get_bootloader_string(const int, void *, const int);
-int get_bootloader_id(const struct driveinfo *);
+
+#include <stdint.h>
+
+void get_bootloader_string(const uint16_t, void *, const int);
+uint16_t get_bootloader_id(const struct driveinfo *);
 #endif /* _BOOTLOADERS_H_ */
